Move Date and LearningBox to components folder
